Add files via upload
The aim of the project was to find the new hub locations in order to satisfy the demand of the future need of hub beneficiaries.
Based in primary schools (state, catholic and independent), community hubs serve as gateways to connect families with each other, with their school and with existing services. Hubs build communities through engagement, early childhood services, English conversation support, and pathways to volunteering, training and employment. There are currently 74 HUBS in Australia where 3,986 families engaged with and 176 individuals found jobs as a result of connecting with a hub.
The proposal of new hubs for schools/LGAs was done by exploratory data analysis, visualization, and modelling specifically applied to school profile in Australia using features of existing school hub and geo spatial data of LGAs. The data were provided by the Community Hub Australia.
The results indicated there were seven different groups of schools and the majority of hubs were in three clusters. The efficiency of each cluster was positive. The three clusters that had larger number of hubs were selected and schools were filtered using the specific attributes they used in the past to implement school hubs. And also, some external data related to socio-economic factors including migrant family background in the LGAs were used in the final selection of LGAs.
We concluded that Casey and Wyndham councils are the potential LGAs to choose schools for new hubs as their population related to non-English speaking is high with high developmentally vulnerable kids’ percentage.
The Audience of the project is Community Hub Australia and so, the study can be continued based on this analysis for future implementation and enhancements of hub quality. It is advisable to include further attributes such as success rates and efficiency of hubs to the hub profiles so that statistically significant schools could be avoided.